refactor(home): flatten hero/ into home/ and drop redundant Home wrapper
Home only ever rendered <ProfileHero>, so the wrapper component and
the nested hero/ folder added an indirection layer with no semantic
value. HomePage now renders ProfileHero directly, and its supporting
pieces (cta-buttons, terminal-animation, mesh-gradient) live flat
under components/home/.
