+++ src/components/home/terminal-animation.tsx
